+++ CHANGELOG.md
@@ -6,6 +6,8 @@ and this project doesn't really care for versioning.
 
 ## [Unreleased]
 
+## [1.9.0] – 2022-01-10
+
 ### Added
 - Implement `LogControl` DBus interface, in order to set the log level with `systemctl service-log-level` (see [GH-27]).
 
@@ -139,7 +141,8 @@ and this project doesn't really care for versioning.
 
 Initial prototype in Python, with support for Toolbox IDEA CE and WebStorm.
 
-[Unreleased]: https://github.com/lunaryorn/gnome-search-providers-jetbrains/compare/v1.8.0...HEAD
+[Unreleased]: https://github.com/lunaryorn/gnome-search-providers-jetbrains/compare/v1.9.0...HEAD
+[1.9.0]: https://github.com/lunaryorn/gnome-search-providers-jetbrains/compare/v1.8.0...v1.9.0
 [1.8.0]: https://github.com/lunaryorn/gnome-search-providers-jetbrains/compare/v1.7.1...v1.8.0
 [1.7.1]: https://github.com/lunaryorn/gnome-search-providers-jetbrains/compare/v1.7.0...v1.7.1
 [1.7.0]: https://github.com/lunaryorn/gnome-search-providers-jetbrains/compare/v1.6.0...v1.7.0